Add "Accept: application/dns-message" as for RFC 8484 Guidance
The DoH client SHOULD include an HTTP Accept request header field to indicate what type of content can be understood in response.
Irrespective of the value of the Accept request header field, the client MUST be prepared to process "application/dns-message" (as described in Section 6) responses but MAY also process other DNS-related media types it receives.